+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add new `StaticTooltip` React component.
 - Fix: Add bottom padding to `.k-SideGrid__content` element.
 - Fix: Fix SearchInput content overflow display.